+++ MslmNetworkManager/Bundle + Extension.swift
