+++ SwiftPackageManager.md
@@ -1,0 +1,14 @@
+# Swift Package Manager for Mslm
+
+## Requirements
+
+- Requires Xcode 14.1 or above
+- Analytics requires clients to add `-ObjC` linker option.
+- See [Package.swift](Package.swift) for supported platform versions.
+
+## Installation
+
+### Installing from Xcode
+
+Add a package by selecting `File` → `Add Packages…` in Xcode’s menu bar.
+
